+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Enhancements
 - GP2-2402 - VFM analytics
 - GP2-2348 - VFM Return copy update
+- GP2-2700 - Export plan/Business risk copy update
+- GP2-2701 - Export plan/Travel plan copy update
+- GP2-2706 - Export plan/Costs and pricing copy update
 
 ### Fixed bugs
 - GP2-3099 - WTE Selector missing